half cleaned code

+++ Node1/CAN_joystick.h
@@ -9,11 +9,12 @@
 #ifndef CAN_JOYSTICK_H_
 #define CAN_JOYSTICK_H_
 
+
 #include "Utilities.h"
 #include "joystick.h"
 #include "CAN.h"
-void send_joystick_dir(int msg_id);
-void send_all_slider(int msg_id);
-void send_all_joy(int msg_id);
+void send_joystick_dir(uint8_t msg_id);
+void send_all_slider(uint8_t msg_id);
+void send_all_joy(uint8_t msg_id);
 
 #endif
